third_party/WebKit/LayoutTests/bluetooth/characteristic/notifications/add-multiple-event-listeners.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let char;
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('heart_rate'))
    .then(service => service.getCharacteristic('heart_rate_measurement'))
    .then(characteristic => {
      char = characteristic;
      return assert_event_fires_after_promise(characteristic,
                                              'startNotifications',
                                              'characteristicvaluechanged',
                                              3 /* add 3 listeners */,
                                              false /* ignore_event_promise_order */);
    })
    .then(() => char.stopNotifications())
    .then(() => assert_no_events(char, 'characteristicvaluechanged'));
}, 'Add multiple event listeners then startNotifications().');
</script>
